requirements.txt
uagents 
uagents_adapter 
python-dotenv